Update substance-form.service.ts
Since js is 0-indexed, we need to start from length-1 and go down. There was old code that was starting from length instead, which is out of range. This still used to work because there were checks inside for whether the value at that invalid index was "truthy". Just switching it to start at the right index is easier though